Seperate class for setup teardown
