+++ test-output/Default suite/Default test.html
@@ -55,11 +55,11 @@ function toggleAllBoxes() {
 <body>
 <h2 align='center'>Default test</h2><table border='1' align="center">
 <tr>
-<td>Tests passed/Failed/Skipped:</td><td>1/0/0</td>
+<td>Tests passed/Failed/Skipped:</td><td>2/0/0</td>
 </tr><tr>
-<td>Started on:</td><td>Sun Apr 14 10:44:38 IST 2024</td>
+<td>Started on:</td><td>Sun Apr 14 16:06:37 IST 2024</td>
 </tr>
-<tr><td>Total time:</td><td>13 seconds (13969 ms)</td>
+<tr><td>Total time:</td><td>9 seconds (9681 ms)</td>
 </tr><tr>
 <td>Included groups:</td><td></td>
 </tr><tr>
@@ -77,8 +77,13 @@ function toggleAllBoxes() {
 <tr>
 <td title='com.mvngitjenkins.SampleTestCase.veriyPageTitle_01()'><b>veriyPageTitle_01</b><br>Test class: com.mvngitjenkins.SampleTestCase</td>
 <td></td>
-<td>4</td>
-<td>com.mvngitjenkins.SampleTestCase@7a419da4</td></tr>
+<td>1</td>
+<td>com.mvngitjenkins.SampleTestCase@4bb33f74</td></tr>
+<tr>
+<td title='com.mvngitjenkins.SampleTestCase.veriyCurrentURL_02()'><b>veriyCurrentURL_02</b><br>Test class: com.mvngitjenkins.SampleTestCase</td>
+<td></td>
+<td>3</td>
+<td>com.mvngitjenkins.SampleTestCase@4bb33f74</td></tr>
 </table><p>
 </body>
 </html>
