+++ test-output/Default suite/Default test.xml
@@ -1,5 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitXMLReporter -->
-<testsuite ignored="0" hostname="DESKTOP-7FQSF63" failures="0" tests="1" name="Default test" time="13.969" errors="0" timestamp="2024-04-14T10:44:51 IST">
-  <testcase classname="com.mvngitjenkins.SampleTestCase" name="veriyPageTitle_01" time="4.093"/>
+<testsuite ignored="0" hostname="DESKTOP-7FQSF63" failures="0" tests="2" name="Default test" time="9.681" errors="0" timestamp="2024-04-14T16:06:47 IST">
+  <testcase classname="com.mvngitjenkins.SampleTestCase" name="veriyPageTitle_01" time="1.931"/>
+  <testcase classname="com.mvngitjenkins.SampleTestCase" name="veriyCurrentURL_02" time="3.124"/>
 </testsuite> <!-- Default test -->
