+++ test-output/junitreports/TEST-com.mvngitjenkins.SampleTestCase.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="DESKTOP-7FQSF63" failures="0" tests="1" name="com.mvngitjenkins.SampleTestCase" time="4.904" errors="0" timestamp="2024-04-14T10:45:58 IST" skipped="0">
-  <testcase classname="com.mvngitjenkins.SampleTestCase" name="veriyPageTitle_01" time="4.904"/>
+<testsuite hostname="DESKTOP-7FQSF63" failures="0" tests="2" name="com.mvngitjenkins.SampleTestCase" time="9.647" errors="0" timestamp="2024-04-14T16:06:47 IST" skipped="0">
+  <testcase classname="com.mvngitjenkins.SampleTestCase" name="veriyPageTitle_01" time="5.375"/>
+  <system-out/>
+  <testcase classname="com.mvngitjenkins.SampleTestCase" name="veriyCurrentURL_02" time="4.272"/>
   <system-out/>
 </testsuite> <!-- com.mvngitjenkins.SampleTestCase -->
