+++ test-output/old/Default suite/classes.html
@@ -11,6 +11,10 @@
 </tr>
 <tr>
 <td>&nbsp;</td>
+<td>veriyCurrentURL_02</td>
+<td>&nbsp;</td></tr>
+<tr>
+<td>&nbsp;</td>
 <td>veriyPageTitle_01</td>
 <td>&nbsp;</td></tr>
 <tr>
