+++ test-output/old/Default suite/methods.html
@@ -1,10 +1,16 @@
 <h2>Methods run, sorted chronologically</h2><h3>&gt;&gt; means before, &lt;&lt; means after</h3><p/><br/><em>Default suite</em><p/><small><i>(Hover the method name to see the test class name)</i></small><p/>
 <table border="1">
 <tr><th>Time</th><th>Delta (ms)</th><th>Suite<br>configuration</th><th>Test<br>configuration</th><th>Class<br>configuration</th><th>Groups<br>configuration</th><th>Method<br>configuration</th><th>Test<br>method</th><th>Thread</th><th>Instances</th></tr>
-<tr bgcolor="8f6f80">  <td>24/04/14 10:44:47</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="SampleTestCase.veriyPageTitle_01()[pri:0, instance:com.mvngitjenkins.SampleTestCase@7a419da4]">veriyPageTitle_01</td> 
-  <td>main@696933920</td>   <td></td> </tr>
-<tr bgcolor="8f6f80">  <td>24/04/14 10:44:51</td>   <td>4097</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;SampleTestCase.tearDown()[pri:0, instance:com.mvngitjenkins.SampleTestCase@7a419da4]">&lt;&lt;tearDown</td> 
-<td>&nbsp;</td>  <td>main@696933920</td>   <td></td> </tr>
-<tr bgcolor="8f6f80">  <td>24/04/14 10:44:38</td>   <td>-9708</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;SampleTestCase.setup()[pri:0, instance:com.mvngitjenkins.SampleTestCase@7a419da4]">&gt;&gt;setup</td> 
-<td>&nbsp;</td>  <td>main@696933920</td>   <td></td> </tr>
+<tr bgcolor="8f6f80">  <td>24/04/14 16:06:40</td>   <td>0</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="SampleTestCase.veriyPageTitle_01()[pri:1, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]">veriyPageTitle_01</td> 
+  <td>main@1754444726</td>   <td></td> </tr>
+<tr bgcolor="677d7f">  <td>24/04/14 16:06:42</td>   <td>2154</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestBase.setup()[pri:0, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]">&gt;&gt;setup</td> 
+<td>&nbsp;</td>  <td>main@1754444726</td>   <td></td> </tr>
+<tr bgcolor="677d7f">  <td>24/04/14 16:06:46</td>   <td>6211</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestBase.tearDown()[pri:0, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@1754444726</td>   <td></td> </tr>
+<tr bgcolor="677d7f">  <td>24/04/14 16:06:42</td>   <td>2154</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&gt;&gt;TestBase.setup()[pri:0, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]">&gt;&gt;setup</td> 
+<td>&nbsp;</td>  <td>main@1754444726</td>   <td></td> </tr>
+<tr bgcolor="677d7f">  <td>24/04/14 16:06:46</td>   <td>6211</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="&lt;&lt;TestBase.tearDown()[pri:0, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]">&lt;&lt;tearDown</td> 
+<td>&nbsp;</td>  <td>main@1754444726</td>   <td></td> </tr>
+<tr bgcolor="8f6f80">  <td>24/04/14 16:06:43</td>   <td>3085</td> <td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td>&nbsp;</td><td title="SampleTestCase.veriyCurrentURL_02()[pri:2, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]">veriyCurrentURL_02</td> 
+  <td>main@1754444726</td>   <td></td> </tr>
 </table>
