+++ test-output/old/Default suite/toc.html
@@ -10,7 +10,7 @@
 <tr valign='top'>
 <td>1 test</td>
 <td><a target='mainFrame' href='classes.html'>1 class</a></td>
-<td>1 method:<br/>
+<td>2 methods:<br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods.html'>chronological</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-alphabetical.html'>alphabetical</a><br/>
 &nbsp;&nbsp;<a target='mainFrame' href='methods-not-run.html'>not run (0)</a></td>
@@ -22,7 +22,7 @@
 </tr></table>
 <table width='100%' class='test-passed'>
 <tr><td>
-<table style='width: 100%'><tr><td valign='top'>Default test (1/0/0)</td><td valign='top' align='right'>
+<table style='width: 100%'><tr><td valign='top'>Default test (2/0/0)</td><td valign='top' align='right'>
   <a href='Default test.html' target='mainFrame'>Results</a>
 </td></tr></table>
 </td></tr><p/>
