+++ test-output/testng-results.xml
@@ -1,25 +1,37 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<testng-results ignored="0" total="1" passed="1" failed="0" skipped="0">
+<testng-results ignored="0" total="2" passed="2" failed="0" skipped="0">
   <reporter-output>
   </reporter-output>
-  <suite started-at="2024-04-14T10:45:53 IST" name="Suite" finished-at="2024-04-14T10:45:58 IST" duration-ms="4933">
+  <suite started-at="2024-04-14T16:06:37 IST" name="Default suite" finished-at="2024-04-14T16:06:47 IST" duration-ms="9681">
     <groups>
     </groups>
-    <test started-at="2024-04-14T10:45:53 IST" name="Test" finished-at="2024-04-14T10:45:58 IST" duration-ms="4933">
+    <test started-at="2024-04-14T16:06:37 IST" name="Default test" finished-at="2024-04-14T16:06:47 IST" duration-ms="9681">
       <class name="com.mvngitjenkins.SampleTestCase">
-        <test-method is-config="true" signature="setup()[pri:0, instance:com.mvngitjenkins.SampleTestCase@7a419da4]" started-at="2024-04-14T10:45:53 IST" name="setup" finished-at="2024-04-14T10:45:56 IST" duration-ms="2151" status="PASS">
+        <test-method is-config="true" signature="setup()[pri:0, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]" started-at="2024-04-14T16:06:37 IST" name="setup" finished-at="2024-04-14T16:06:40 IST" duration-ms="3260" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- setup -->
-        <test-method signature="veriyPageTitle_01()[pri:0, instance:com.mvngitjenkins.SampleTestCase@7a419da4]" started-at="2024-04-14T10:45:56 IST" name="veriyPageTitle_01" finished-at="2024-04-14T10:45:58 IST" duration-ms="2607" status="PASS">
+        <test-method signature="veriyPageTitle_01()[pri:1, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]" started-at="2024-04-14T16:06:40 IST" name="veriyPageTitle_01" finished-at="2024-04-14T16:06:42 IST" duration-ms="1931" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- veriyPageTitle_01 -->
-        <test-method is-config="true" signature="tearDown()[pri:0, instance:com.mvngitjenkins.SampleTestCase@7a419da4]" started-at="2024-04-14T10:45:58 IST" name="tearDown" finished-at="2024-04-14T10:45:58 IST" duration-ms="146" status="PASS">
+        <test-method is-config="true" signature="tearDown()[pri:0, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]" started-at="2024-04-14T16:06:42 IST" name="tearDown" finished-at="2024-04-14T16:06:42 IST" duration-ms="217" status="PASS">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- tearDown -->
+        <test-method is-config="true" signature="setup()[pri:0, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]" started-at="2024-04-14T16:06:42 IST" name="setup" finished-at="2024-04-14T16:06:43 IST" duration-ms="931" status="PASS">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- setup -->
+        <test-method signature="veriyCurrentURL_02()[pri:2, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]" started-at="2024-04-14T16:06:43 IST" name="veriyCurrentURL_02" finished-at="2024-04-14T16:06:46 IST" duration-ms="3124" status="PASS">
+          <reporter-output>
+          </reporter-output>
+        </test-method> <!-- veriyCurrentURL_02 -->
+        <test-method is-config="true" signature="tearDown()[pri:0, instance:com.mvngitjenkins.SampleTestCase@4bb33f74]" started-at="2024-04-14T16:06:46 IST" name="tearDown" finished-at="2024-04-14T16:06:47 IST" duration-ms="184" status="PASS">
           <reporter-output>
           </reporter-output>
         </test-method> <!-- tearDown -->
       </class> <!-- com.mvngitjenkins.SampleTestCase -->
-    </test> <!-- Test -->
-  </suite> <!-- Suite -->
+    </test> <!-- Default test -->
+  </suite> <!-- Default suite -->
 </testng-results>
